Replace prec:IriOfSomething syntax with Rule syntaxes
:name prec:IRIOfRelationship [
  prec:relationshipLabel "name" ;
  prec:modelAs :someModel
  prec:aSubstitutionTerm :someBrandNewIRI
]

Was a strange syntax because:
- Some information were as the subject of prec:IRIOfRelationship
while other were at the object position of the blank node at the
object position (lack of symetry)
- It was possible for :someModel to ignore the :name parameter
(information loss)
- the provided IRI and the substitution terms had overlapping roles

This commit removes the possibility to use blank node in the object
position of prec:IRIOfRelationship.

Instead, users are now required to write rules, for example:
[] a prec:RelationshipRule ;
  prec:relationshipLabel "Name" ;
  prec:relationshipIRI :name ;
  prec:modelAs :someModel ;
  prec:aSubstitutionTerm :someBrandNewIRI .

relationshipIRI is a now substitution term that takes the roles of the
subject of prec:IRIOfRelationship).

This syntax also looks more like SHACL, which is probably good
design as while SHACL is schema for RDF graphs, the filter predicates
of PREC can be seen as a schema for Property Graphs written in RDF.